+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 # Music #
 + The lyrics command's response's title now is the title of the song instead of the query. However, the query is visible as the footer.
 + Custom error classes.
++ Loop command can now loop playlists.
 
 # Util #
 + The clear command now does not purge pinned messages.